Reformatted code
Fixed some errors in the existing code. Also changed some code to get
it work as it should.
